% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slotmachine.R
\name{score}
\alias{score}
\title{score}
\usage{
score(symbols)
}
\description{
get score based in symbols generated by get_symbols
}
\examples{
\dontrun{
set.seed(24)
one_play <- get_symbols()
score(one_play)
}
}
